services/sports-telemetry/requirements.txt
google-cloud-pubsub==2.19.0
